valve_pi: floatless

+++ valve.h
@@ -42,7 +42,7 @@ struct s_valve_pi_priv {
 		timekeep_t last_time;	///< last time the PI controller algorithm was run
 		timekeep_t Tc;		///< closed loop time constant
 		temp_t prev_out;	///< previous run output temperature
-		float Kp_t;		///< Kp time factor: Kp = Kp_t / K, K process gain, Kp proportional coefficient
+		uint_fast32_t Kp_t;	///< Kp time factor: Kp = Kp_t / K, K process gain, Kp proportional coefficient
 		int_fast32_t db_acc;	///< deadband accumulator. Needed to integrate when valve is not actuated despite request.
 	} run;		///< private runtime (internally handled)
 };
